pyproject.toml
[tool.poetry]
name = "apitable-scripts"
version = "0.1.0"
description = "Call apitable openapi"
authors = ["samzong.lu <samzong.lu@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "apitable_scripts"}]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.22.0"
fastapi = "^0.98.0"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
